+++ conanfile.py
@@ -11,7 +11,6 @@ class VeriluaConan(ConanFile):
         # in order to get shared libraries for libgmp.
         "slang/10.0",
         "sol2/3.3.1",
-        "argparse/3.1",
         "inja/3.4.0",
         "libassert/2.1.4",
         "fmt/12.1.0",